moved the merge pipeline stuff as well
