+++ quickwit/quickwit-indexing/src/actors/log_pipeline/merge_executor.rs
