+++ quickwit/quickwit-indexing/src/actors/log_pipeline/merge_scheduler_service.rs
