+++ quickwit/quickwit-indexing/src/actors/log_pipeline/merge_split_downloader.rs
