+++ quickwit/quickwit-indexing/src/actors/log_pipeline/mod.rs
@@ -19,11 +19,19 @@
 //! ```text
 //! Source → DocProcessor → Indexer → IndexSerializer → Packager → Uploader → Publisher
 //! ```
+//!
+//! It also contains the merge pipeline actors, which are only used by the
+//! logs/traces pipeline (the metrics pipeline does not perform merges).
 
 mod cooperative_indexing;
 mod doc_processor;
 mod index_serializer;
 mod indexer;
+mod merge_executor;
+pub(crate) mod merge_pipeline;
+mod merge_planner;
+mod merge_scheduler_service;
+mod merge_split_downloader;
 mod packager;
 mod pipeline;
 #[cfg(feature = "vrl")]
@@ -32,5 +40,10 @@ mod vrl_processing;
 pub use doc_processor::{DocProcessor, DocProcessorCounters};
 pub use index_serializer::IndexSerializer;
 pub use indexer::{Indexer, IndexerCounters};
+pub use merge_executor::{MergeExecutor, combine_partition_ids, merge_split_attrs};
+pub use merge_pipeline::{FinishPendingMergesAndShutdownPipeline, MergePipeline, MergePipelineParams};
+pub(crate) use merge_planner::{MergePlanner, RunFinalizeMergePolicyAndQuit};
+pub use merge_scheduler_service::{MergePermit, MergeSchedulerService, schedule_merge};
+pub use merge_split_downloader::MergeSplitDownloader;
 pub use packager::Packager;
 pub use pipeline::{IndexingPipeline, IndexingPipelineParams};
